semana10/dia6/perudemy/backend/requirements.txt
django==3.2
djangorestframework
mysqlclient
cloudinary
django-cors-headers
djangorestframework-simplejwt

